requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.0.0
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.0
Django==5.0.3
django-browser-reload==1.12.1
django-environ==0.11.2
django-tailwind==3.8.0
factory-boy==3.3.0
Faker==24.4.0
gunicorn==21.2.0
heroicons==2.6.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
packaging==24.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.10.0
whitenoise==6.6.0
zope.interface==6.2
